move ticket to new module
This has the advantage that the provider itself can not construct the
ticket directly and thus is forced to uphold the invariant that the
address list must not be empty.
